requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
psycopg2-binary
pydantic
python-dotenv
python-multipart
pandas
jwt
apache-airflow
faker